DES-143: Define versions carefully within ConfigureServices
The Swashbuckle CLI only runs the ConfigureServices startup command,
previously the Configure Services command had a dependency on Configure
to have run.

Now the versions are configured as an array in ConfigureServices in
order to also set up Swashbuckle. And the Configure can use the versions
from the api service.